refactor: :test_tube: cleanup test sutes and dashboard components
