+++ packages/bot-web-ui/src/pages/dashboard/bot-list/__tests__/stop-bot-modal.spec.tsx
