+++ packages/bot-web-ui/src/pages/dashboard/bot-list/delete-dialog/__tests__/delete-dialog.spec.tsx
