+++ packages/bot-web-ui/src/pages/dashboard/bot-list/delete-dialog/delete-dialog.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import '../../../../styles/mixins.scss';
 
 .dc-dialog__delete-strategy--delete {
     .dc-dialog {
