+++ packages/bot-web-ui/src/pages/dashboard/bot-list/delete-dialog/index.ts
@@ -1,0 +1,3 @@
+import DeleteDialog from './delete-dialog';
+
+export default DeleteDialog;
