+++ packages/bot-web-ui/src/pages/dashboard/bot-list/google-drive/index.ts
@@ -1,0 +1,3 @@
+import GoogleDrive from './google-drive';
+
+export default GoogleDrive;
