+++ packages/bot-web-ui/src/pages/dashboard/bot-list/index.scss
@@ -1,7 +1,4 @@
 @import '../../../styles/mixins.scss';
-@import './google-drive.scss';
-@import './delete-dialog.scss';
-@import './save-modal.scss';
 
 .load-strategy {
     &__wrapper {
