+++ packages/bot-web-ui/src/pages/dashboard/bot-list/save-modal/index.ts
@@ -1,0 +1,3 @@
+import SaveModal from './save-modal';
+
+export default SaveModal;
