+++ packages/bot-web-ui/src/pages/dashboard/bot-list/save-modal/save-modal.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import '../../../../styles/mixins.scss';
 
 .dc-modal {
     $save-modal: 'modal--save';
