--- packages/bot-web-ui/src/pages/dashboard/load-bot-preview/index.ts
@@ -1,4 +1,0 @@
-import LoadBotPreview from './local';
-import './index.scss';
-
-export default LoadBotPreview;
